Create File: "theme-structure-schema.json". File contains an incomplete schema.
The schema contained within the file at its initial commit, is an incomplete schema that needs much more work done before it complete. The schema is used by the VS Code extension 'Theme Dojo', which at the time of this commit is not yet published to the VS Code marketplace.